changes for home

+++ app/src/main/java/com/example/restopass/service/MembershipService.kt
@@ -52,7 +52,7 @@ object MembershipService {
                 it.description,
                 it.img,
                 it.visits,
-                it.price,
+                it.price.toInt(),
                 restaurantsWithAccordingDishes
             )
         }
